requirements.txt
numpy
PILLOW
transformers
mlflow
pandas
torchvision
flask